do not use git shallow clone/fetch
jenkins started failing to fetch on pre-existing lsst/lsstsw clones this
morning.  This is first time this failure mode has been observed.

This jenkins jira issue suggests that the problem is specific to the
version of git shipped with el7 when doing a shallow clone/fetch over
http[s]:

https://issues.jenkins-ci.org/browse/JENKINS-37229

Current jenkins master:

    $ cat /etc/redhat-release
    CentOS Linux release 7.1.1503 (Core)

    $ git --version
    git version 1.8.3.1

    $ git fetch --tags --progress https://github.com/lsst/lsstsw.git +refs/heads/*:refs/remotes/origin/* --depth=1 --verbose
    POST git-upload-pack (gzip 16429 to 5558 bytes)
    POST git-upload-pack (gzip 17233 to 5661 bytes)
    POST git-upload-pack (gzip 18033 to 5754 bytes)
    POST git-upload-pack (gzip 19633 to 5938 bytes)
    POST git-upload-pack (gzip 22833 to 6302 bytes)
    fatal: git fetch-pack: expected shallow list