rm unneeded usage of --storage-opt
Specifying a non-default container rootfs size only works with overlay2
on xfs when pquota (xfs project quotas) are enabled.  This is exposing a
docker configuration problem, that will be fixed separately. However, it
does appear that a non-default rootfs is no long required where this
optional was being used and is safe to remove.