do not create EupsBuildDir as a side effect of cleaning up
Previously, the side effects of using the dir() step would create this
path if it did not already exist.  As the path is constructed using the
what was intended to be the `current` "symlink", this would result in a
dir existing in place of the symlink, which would cause re-runs of
`newinstall.sh` to fail.