Use hash for the build directory
We have problems with the path length of the build directory growing and causing issue with path length limits.  By using a hash of the nominal path, we can keep the build path a fixed length while still providing the necessary information to the build steps.  We use SHA1 as the hash algorithm and truncate to 10 characters to save enough space.