Add rbClassifier_data to ap_verify
Original Issue-
Add the DC2 real bogus trained model to the new CI dataset. In addition
to adding the data to the git-lfs repo, add the trained model to the
ingest step of ap_verify. This will require defining the trained model
for the butler, and it is suggested that we treat the model the same way
as we treat our "curated calibration" datasets.

Change Made-
1. Added a block of code that will-
    -Create a new directory for rbClassifier_data.
    -Pull the rbClassifier_data repo into the newly created directory.
2. Modified the bash script inside the runApVerify method to run the
setup inside the rbClassifier_data directory.

For Future-
After Butler implementation is in place, remove the rbClassifier_data
block and the corresponding variables from ap_verify.groovy.