Record Jenkins branch refs as Sasquatch metadata.
In SQuaSH, the ci_refs field was derived from the $refs environment
variable, which was set in scipipe/ap_verify but for some reason
didn't end up in the SQuaSH output.

This version moves the definition of $refs (now $JOB_REFS) inside
runVerifyToSasquatch, avoiding an implicit dependency on the caller.
The refs are an optional argument for compatibility with
sqre/verify_drp_metrics.groovy, which did not use the original $refs.